+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.4.0
 * `produce_or_load` now allows the possibility of not loading the file
+* New function `struct2dict` that converts a struct to a dictionary (for saving)
 # 0.3.0
 * Added `test` as a directory of the default project (#43)
 * Added `tmpsave` functionality: save the result of `dict_list` in temporary files and conveniently work with sequential clusters (#45)